int to long all ID and user to patient many to one relation

+++ backend/teleconsulting/src/main/java/com/had/teleconsulting/teleconsulting/Bean/Appointment.java
@@ -17,7 +17,7 @@ public class Appointment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false,name = "appointmentID")
-    int appointmentID;
+    long appointmentID;
 
     @Column(nullable = false,name = "appointmentOpdType")
     String appointmentOpdType;
