+++ backend/teleconsulting/src/main/java/com/had/teleconsulting/teleconsulting/Bean/DoctorDetails.java
@@ -16,7 +16,7 @@ public class DoctorDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable=false,name="doctorID")
-    int doctorID;
+    long doctorID;
 
     @Column(nullable=false,name="doctorFirstName")
     String doctorFirstName;
